feat: add ModelPopulater class for importing and exporting model data from Excel
